Create README.md
The telecommunication dataset is a classification problem aimed at segmenting its customers based on service usage patterns, categorizing the customer into four groups ie basic, e-service, plus service and total service. The dmographoc data will be used to predict group by customizing offers for individual prospective customers. The objective of this assignment is to build a classifier to predict the class of unknown cases. We will be using the KNN to analyze this dataset.